requirements.txt
sentence_transformers==2.2.0
python-telegram-bot